Move Posto dataclass from models.data to models.deolhonafila
