+++ filometro/models/deolhonafila.py
@@ -2,6 +2,9 @@
 
 import requests
 
+from dataclasses import dataclass
+from dataclasses import field
+
 
 ENDPOINT = 'https://deolhonafila.prefeitura.sp.gov.br/processadores/dados.php'
 
@@ -28,6 +31,26 @@ HEADERS = {
 }
 
 PAYLOAD = {'dados': 'dados'}
+
+
+@dataclass
+class Posto():
+    equipamento: str = field(repr=False)
+    endereco: str = field(repr=False)
+    distrito: str
+    crs: str = field(repr=False)
+    astrazeneca: str = field(repr=False)
+    coronavac: str = field(repr=False)
+    pfizer: str = field(repr=False)
+    intercambialidade: str = field(repr=False)
+    status_fila: str
+    indice_fila: str = field(repr=False)
+    data_hora: str
+    tipo_posto: str = field(repr=False)
+    id_crs: str = field(repr=False)
+    id_distrito: str = field(repr=False)
+    id_tb_unidades: str = field(repr=False)
+    id_tipo_posto: str = field(repr=False)
 
 
 class APIDeOlhoNaFila():
